+++ src/main/java/com/excilys/cdb/persistence/ComputerDAOInterface.java
@@ -12,15 +12,17 @@ public interface ComputerDAOInterface<T> {
 
 	public abstract void delete(Long id) ;
 
-	
+
 	public abstract void update(T obj) ;
 
-	
+
 	public abstract Optional<Computer> find(Long id) ;
 
-	
+
 	public abstract List<T> findAll(String name, int page, int size) ;
-	int count(String name) ;
+
+
+	long count(String name) ;
 
 	void deleteByCompany(Long companyId);
